# swagger.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Online Library API
  description: A simple API to do CRUD on users, books, orders resources in online library
servers:
  - url: http://localhost:5000/
...
paths:
  /users:
    get:
      summary: Lists all the users
      tags: [User]
      responses:
        "200":
          description: The list of users.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    post:
      summary: Creates an user Object .
      tags: [User]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "201":
          description: Created Successfully.
...
  /books:
    get:
      summary: Lists all the Books
      tags: [Book]
      responses:
        "200":
          description: The list of books.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
    post:
      summary: Creates a Book Object.
      tags: [Book]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Book"
      responses:
        "201":
          description: Created Successfully.
...
  /orders:
    get:
      summary: Lists all the Orders
      tags: [Order]
      responses:
        "200":
          description: The list of orders.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    post:
      summary: Creates an Order Object.
      tags: [Order]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Order"
      responses:
        "201":
          description: Created Successfully.
...
  /search/books:
    get:
      tags: [Search]
      parameters:
        - in: header
          name: x-access-token
          schema: 
            type: string
          required: true 
        - in: query
          name: title
          schema:
            type: string
          description: The title of the book.
        - in: query
          name: author
          schema:
            type: string
          description: The author of the book.
        - in: query
          name: ISBN
          schema:
            type: string
          description: The ISBN of the book.
        - in: query
          name: publication
          schema:
            type: string
          description: The publication of the book.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Search"
  /search/orders:
    get:
      tags: [Search]
      parameters:
        - in: header
          name: x-access-token
          schema: 
            type: string
          required: true 
        - in: query
          name: username
          schema:
            type: string
          description: The username who ordered the book.
        - in: query
          name: return
          schema:
            type: boolean
          description: Whether the book is returned or not.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Search"
components:
  schemas:
    User:
      type: object
      required:
        - username
        - password
        - isAdmin
      properties:
        id:
          type: integer
          description: The auto-generated id of the student.
        username:
          type: string
          description: The name of the student.
        password:
          type: string
          description: The roll number of the student
        isAdmin:
          type: boolean
          description: Checks whether the user is Admin or not 
      example:
        username: bada
        password: b123
        isAdmin: false
    Book:
      type: object
      required:
        - title
        - author
        - ISBN
        - publication
        - link
        - addedBy
      properties:
        id:
          type: integer
          description: The auto-generated id of the book.
        title:
          type: string
          description: The title of the book.
        author:
          type: string
          description: The author of the book.
        ISBN:
          type: string
          description: The ISBN number of the book.
        publication:
          type: string
          description: The year of publication of the book.
        link:
          type: string
          description: The link of the image of the book.
        addedBy:
          type: string
          description: The admin name who added the book.
      example:
        title: crazy kids
        author: bada
        ISBN: 978-3-16-148410-0
        publication: 2001
        link: https://d1w7fb2mkkr3kw.cloudfront.net/assets/images/book/lrg/9781/4088/9781408894996.jpg
        addedBy: manoj
    Order:
      type: object
      required:
        - date
        - username
        - return
        - bookID
      properties:
        id:
          type: integer
          description: The auto-generated id of the respective order.
        username:
          type: string
          decription: the name of user who ordered the book.
        date:
          type: string
          description: the time a book ordered by user.
        return:
          type: boolean
          description: whether returned in time or not.
        bookId:
          type: string
          description: the ID of the book ordered.
      example:
        date: Wed, 14 Jun 2017 07:00:00 GMT
        username: bada
        return: true
        bookId: 624703f88534ab047cc54ecd
...
    Search:
      type: object